Исправлено: 1. При редактировании операции теперь подставляется вся информация о транзакции. 2. К кнопке удалить прикручена форма 3. Изменено поведение алертов на сттранице с транзакциями (теперь они отображаются в виде модального окна) 4. Из шаблона Base удален ненужный dropdown

+++ finance/accounting/views.py
@@ -116,7 +116,9 @@ class TransactionUpdateView(View):
 
 
 class TransactionDeleteView(View):
-    def delete(self, request, id: int, *args, **kwargs):
-        Transaction.objects.get(id=id).delete()
-        messages.success(request, "Транзакция удалена")
+    def post(self, request, id: int, *args, **kwargs):
+        if request.POST.get('_method') == 'DELETE':
+            Transaction.objects.get(id=id).delete()
+            messages.success(request, "Транзакция удалена")
+            return redirect("accounting:transactions")
         return redirect("accounting:transactions")
